LPS-90300 Move display class to the display package
