+++ modules/apps/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/display/ScopeDisplay.java
@@ -12,9 +12,7 @@
  * details.
  */
 
-package com.liferay.user.associated.data.web.internal.portlet.action;
-
-import com.liferay.user.associated.data.web.internal.display.UADApplicationSummaryDisplay;
+package com.liferay.user.associated.data.web.internal.display;
 
 import java.util.List;
