+++ modules/apps/user-associated-data/user-associated-data-web/src/main/resources/META-INF/resources/init.jsp
@@ -53,6 +53,7 @@ page import="com.liferay.taglib.search.ResultRow" %><%@
 page import="com.liferay.user.associated.data.display.UADDisplay" %><%@
 page import="com.liferay.user.associated.data.web.internal.constants.UADConstants" %><%@
 page import="com.liferay.user.associated.data.web.internal.constants.UADWebKeys" %><%@
+page import="com.liferay.user.associated.data.web.internal.display.ScopeDisplay" %><%@
 page import="com.liferay.user.associated.data.web.internal.display.UADApplicationExportDisplay" %><%@
 page import="com.liferay.user.associated.data.web.internal.display.UADApplicationSummaryDisplay" %><%@
 page import="com.liferay.user.associated.data.web.internal.display.UADEntity" %><%@
@@ -62,7 +63,6 @@ page import="com.liferay.user.associated.data.web.internal.display.ViewUADEntiti
 page import="com.liferay.user.associated.data.web.internal.display.context.UADExportProcessDisplayContext" %><%@
 page import="com.liferay.user.associated.data.web.internal.display.context.UADExportProcessManagementToolbarDisplayContext" %><%@
 page import="com.liferay.user.associated.data.web.internal.display.context.ViewUADEntitiesManagementToolbarDisplayContext" %><%@
-page import="com.liferay.user.associated.data.web.internal.portlet.action.ScopeDisplay" %><%@
 page import="com.liferay.user.associated.data.web.internal.search.UADApplicationExportDisplayChecker" %><%@
 page import="com.liferay.user.associated.data.web.internal.search.UADExportProcessResultRowSplitter" %><%@
 page import="com.liferay.user.associated.data.web.internal.util.SafeDisplayValueUtil" %><%@
